fix(portal-v3): reuse existing action detail route
