+++ web/app/(portal)/teams/[teamId]/apps/[appId]/world-id-actions/[actionId]/danger/page.tsx
@@ -1,12 +1,22 @@
 import { pickPortalVersion } from "@/lib/feature-flags/portal-v3/activation";
+import { urls } from "@/lib/urls";
 import { WorldIdActionIdDangerPage } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Danger/page";
-import { WorldIdActionIdDangerPage as WorldIdActionIdDangerPageV3 } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Danger/page";
+import { redirect } from "next/navigation";
 
 export default async function Page(props: {
   params: Promise<Record<string, string>>;
 }) {
   return pickPortalVersion(
-    () => <WorldIdActionIdDangerPageV3 params={props.params} />,
+    async () => {
+      const params = await props.params;
+      redirect(
+        urls.worldIdAction({
+          team_id: params.teamId,
+          app_id: params.appId,
+          action_id: params.actionId,
+        }),
+      );
+    },
     () => <WorldIdActionIdDangerPage params={props.params} />,
   );
 }
