+++ web/app/(portal)/teams/[teamId]/apps/[appId]/world-id-actions/[actionId]/layout.tsx
@@ -1,6 +1,5 @@
 import { pickPortalVersion } from "@/lib/feature-flags/portal-v3/activation";
 import { WorldIdActionIdLayout } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/layout";
-import { WorldIdActionIdLayout as WorldIdActionIdLayoutV3 } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/layout";
 import { ReactNode } from "react";
 
 export default async function Layout(props: {
@@ -8,11 +7,7 @@ export default async function Layout(props: {
   children: ReactNode;
 }) {
   return pickPortalVersion(
-    () => (
-      <WorldIdActionIdLayoutV3 params={props.params}>
-        {props.children}
-      </WorldIdActionIdLayoutV3>
-    ),
+    () => <>{props.children}</>,
     () => (
       <WorldIdActionIdLayout params={props.params}>
         {props.children}
