+++ web/app/(portal)/teams/[teamId]/apps/[appId]/world-id-actions/[actionId]/page.tsx
@@ -1,12 +1,22 @@
 import { pickPortalVersion } from "@/lib/feature-flags/portal-v3/activation";
+import { getIsUserAllowedToUpdateApp } from "@/lib/permissions";
 import { WorldIdActionIdPage } from "@/scenes/Portal/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/page";
-import { WorldIdActionIdPage as WorldIdActionIdPageV3 } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/page";
+import { WorldIdActionDetailPage } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldId/Actions/ActionId/page";
 
 export default async function Page(props: {
   params: Promise<Record<string, string>>;
 }) {
   return pickPortalVersion(
-    () => <WorldIdActionIdPageV3 params={props.params} />,
+    async () => {
+      const params = await props.params;
+
+      return (
+        <WorldIdActionDetailPage
+          params={params}
+          canDelete={await getIsUserAllowedToUpdateApp(params.appId)}
+        />
+      );
+    },
     () => <WorldIdActionIdPage params={props.params} />,
   );
 }
