+++ web/app/(portal)/teams/[teamId]/apps/[appId]/world-id-actions/layout.tsx
@@ -1,6 +1,4 @@
-import { pickPortalVersion } from "@/lib/feature-flags/portal-v3/activation";
 import { urls } from "@/lib/urls";
-import { WorldIdLayout } from "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldId/layout";
 import { fetchAppEnvCached } from "@/scenes/common/Teams/TeamId/Apps/AppId/layout/server/fetch-app-env";
 import { redirect } from "next/navigation";
 import { ReactNode } from "react";
@@ -14,7 +12,7 @@ type Props = {
 export default async function Layout(props: Props) {
   const params = await props.params;
   const { children } = props;
-  const { app, action } = await fetchAppEnvCached(params.appId);
+  const { app } = await fetchAppEnvCached(params.appId);
 
   if (!app?.[0]?.rp_registration?.length) {
     redirect(
@@ -22,12 +20,5 @@ export default async function Layout(props: Props) {
     );
   }
 
-  return pickPortalVersion(
-    () => (
-      <WorldIdLayout hasLegacyActions={action.length > 0}>
-        {children}
-      </WorldIdLayout>
-    ),
-    () => <>{children}</>,
-  );
+  return <>{children}</>;
 }
