+++ web/tests/unit/pv3-r-wia-settings.test.tsx
@@ -2,8 +2,16 @@
 import "@testing-library/jest-dom";
 import { render, screen } from "@testing-library/react";
 import React from "react";
+
+const mockRedirect = jest.fn();
+let mockPortalV3Enabled = true;
+
 jest.mock("@/lib/feature-flags/portal-v3/activation", () => ({
-  pickPortalVersion: async (v3: () => unknown) => v3(),
+  pickPortalVersion: async (v3: () => unknown, v2: () => unknown) =>
+    mockPortalV3Enabled ? v3() : v2(),
+}));
+jest.mock("next/navigation", () => ({
+  redirect: (...args: unknown[]) => mockRedirect(...args),
 }));
 jest.mock(
   "@/scenes/Portal/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Settings/page",
@@ -11,19 +19,30 @@ jest.mock(
     WorldIdActionIdSettingsPage: () => <div data-testid="v2-wia-settings" />,
   }),
 );
-jest.mock(
-  "@/scenes/PortalV3/Teams/TeamId/Apps/AppId/WorldIdActions/ActionId/Settings/page",
-  () => ({
-    WorldIdActionIdSettingsPage: () => <div data-testid="v3-wia-settings" />,
-  }),
-);
 import RoutePage from "../../app/(portal)/teams/[teamId]/apps/[appId]/world-id-actions/[actionId]/settings/page";
-it("renders v3 wia-settings", async () => {
-  render(
-    await RoutePage({
-      params: Promise.resolve({ teamId: "t", appId: "a", actionId: "x" }),
-    }),
+
+const props = () => ({
+  params: Promise.resolve({ teamId: "t", appId: "a", actionId: "x" }),
+});
+
+beforeEach(() => {
+  jest.clearAllMocks();
+  mockPortalV3Enabled = true;
+});
+
+it("returns v3 settings links to the canonical detail", async () => {
+  await RoutePage(props());
+
+  expect(mockRedirect).toHaveBeenCalledWith(
+    "/teams/t/apps/a/world-id-actions/x",
   );
-  expect(screen.getByTestId("v3-wia-settings")).toBeInTheDocument();
-  expect(screen.queryByTestId("v2-wia-settings")).not.toBeInTheDocument();
+});
+
+it("keeps the settings page for v2", async () => {
+  mockPortalV3Enabled = false;
+
+  render(await RoutePage(props()));
+
+  expect(screen.getByTestId("v2-wia-settings")).toBeInTheDocument();
+  expect(mockRedirect).not.toHaveBeenCalled();
 });
